spite config.ts into 2 files
